Fixes #3: Indexes the catalog locally
Motivation
----------
Git cloning is an expensive operation and it's performed while indexing the booster-catalog repository.
This introduces slower boot times and also it's highly-dependable on the Github network availability.

Modifications
-------------
The build process creates a ZIP file from the "next" and the "openshift-online-free" branches and stores inside the resulting JAR,
using the catalog ref as part of the ZIP file name.

Result
------
If the desired ZIP exists, the indexer just extracts it in a temporary directory, avoiding any network roundtrips, otherwise the original behavior is performed.