fix(booster): `content()` will now retry
The `content()` method used to do its work only once, regardless
if the result was correct or resulted in an exception. Now it will
retry if the previous result was an exception.
For connection problems or other temporary problems this will result
in things getting fixed automatically _at some point_. For permanent
problems (like a configuration error) this will result in unending
retries each time someone tries to get the contents of the booster.