feat(indexing): `reindex()` is now "safe"
Before calling `reindex()` would immediately clear the list of boosters
and start the indexing procedure anew, just like it would be done at
startup. But this means that a use might actually see an incomplete list
while 5 minutes ago they saw a complete one, which is confusing. So right
now the indexing is done, but the new list of boosters isn't swapped in
until the indexing has completed.